preformatting system finished
prefixes are only visable/used on the core, its not part of the syntax,
this is done to parse the code vector faster.
each thing now has a prefix so the wordtype is faster known, and string
searches. it improved with another 11%

at the core rustfunction will get prefix 1func()
class functions be \class.func()
struct be |
ncfuncs be 2

at this point the refactored version is 4.6x faster